Module Kusariyarou

This library is based on "Amortized Efficiency of a Path Retrieval Data Structure" by G.F. Italiano, probably with O(log n) slowdown due to the use of Map. This is intended to be used in the cases where graphs are relatively small.

The amortized bound in general does not hold if one forks a graph (for example, adding different edges to the same graph to generate different graphs). However, due to its simplicity, the algorithm should be fine for small graphs.

module type S = sig ... end

The interface of a graph.

module Make : functor (V : Stdlib.Map.OrderedType) -> S with type vertex = V.t

The functor that takes a vertex module and outputs a graph module.