Module type Kusariyarou.S

The interface of a graph.

type t

The type of directed graphs.

type vertex

The type of vertices.

val empty : t

The empty graph.

val mem_vertex : vertex -> t -> bool

mem_vertex v g tests whether the vertex v is in a graph g.

val add_vertex : vertex -> t -> t

add_vertex v g adds the vertex v into the graph g.

val add_edge : vertex -> vertex -> t -> t

add_edge u v g adds an edge from u to v into the graph g. The two vertices must be already in the graph, or Not_found will be raised.

val reachable : vertex -> vertex -> t -> bool

reachable u v g tests whether there is a path from u to v in the transitive and reflexive closure of g. This means reachability is reflexive even if no loop was explicitly added to the graph. If either vertex is not in the graph, Not_found will be raised.